+++ sdk/appconfiguration/azure-appconfiguration/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fixed async `update_sync_token` to use async/await keywords
 
 ### Other Changes
+- Set the default `api_version` to "1.0".
 
 ## 1.4.0 (2022-02-13)